% Auto-generated: do not edit by hand
\name{dashTimeAgo}

\alias{dashTimeAgo}

\title{DashTimeAgo component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
dashTimeAgo(id=NULL, date=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{date}{Character. A label that will be printed when this component is rendered.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

